rename some files and methods in api folder
