+++ src/api/doiRequestApi.ts
@@ -7,7 +7,7 @@ export enum DoiRequestApiPaths {
   DOI_REQUEST = '/doi-request',
 }
 
-export const getPublicationsWithPendingDoiRequest = async (role: RoleName, cancelToken?: CancelToken) =>
+export const getRegistrationsWithPendingDoiRequest = async (role: RoleName, cancelToken?: CancelToken) =>
   await authenticatedApiRequest<Publication[]>({
     url: `${DoiRequestApiPaths.DOI_REQUEST}?role=${role}`,
     cancelToken,
