+++ src/utils/hooks/useFetchPublication.ts
@@ -5,7 +5,7 @@ import Axios from 'axios';
 import { setNotification } from '../../redux/actions/notificationActions';
 import { NotificationVariant } from '../../types/notification.types';
 import { Publication } from '../../types/publication.types';
-import { getPublication } from '../../api/publicationApi';
+import { getRegistration } from '../../api/registrationApi';
 import { RootStore } from '../../redux/reducers/rootReducer';
 
 const useFetchPublication = (
@@ -24,7 +24,7 @@ const useFetchPublication = (
     const cancelSource = Axios.CancelToken.source();
 
     const fetchPublication = async () => {
-      const publication = await getPublication(identifier!, cancelSource.token);
+      const publication = await getRegistration(identifier!, cancelSource.token);
       if (publication) {
         if (publication.error) {
           dispatch(setNotification(publication.error, NotificationVariant.Error));
