+++ src/utils/hooks/useSearchPublications.ts
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
-import { search } from '../../api/publicationApi';
+import { search } from '../../api/registrationApi';
 import { setNotification } from '../../redux/actions/notificationActions';
 import { NotificationVariant } from '../../types/notification.types';
 import { SearchResult } from '../../types/search.types';
